backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.3
shap==0.43.0
fpdf2==2.7.0
python-multipart==0.0.6
aiofiles==23.2.1